+++ packages/extension/src/view/devtools/pages/siteBoundaries/siteBoundaries.tsx
@@ -17,55 +17,13 @@
  * External Dependencies
  */
 import React from 'react';
-import {
-  LandingPageContainer,
-  SIDEBAR_ITEMS_KEYS,
-} from '@google-psat/design-system';
+import { LandingPageContainer } from '@google-psat/design-system';
 import { I18n } from '@google-psat/i18n';
 
-const content = [
-  {
-    title: () => 'Cookies',
-    description: () =>
-      'Insights into the distribution and behavior of cookies on web pages while users navigate across sites during browsing sessions.',
-    url: 'https://developers.google.com/privacy-sandbox/cookies',
-    storyUrl:
-      'https://privacysandbox-stories.com/web-stories/chrome-shifts-to-user-choice-for-third-party-cookies/',
-    sidebarItemKey: SIDEBAR_ITEMS_KEYS.COOKIES,
-  },
-  {
-    title: () => I18n.getMessage('chips'),
-    description: () => I18n.getMessage('chipsDescription'),
-    url: 'https://developers.google.com/privacy-sandbox/3pcd/chips',
-    storyUrl: 'https://privacysandbox-stories.com/web-stories/chips/',
-    sidebarItemKey: SIDEBAR_ITEMS_KEYS.CHIPS,
-  },
-  {
-    title: () => I18n.getMessage('storageAccessAPI'),
-    description: () => I18n.getMessage('storageAccessAPIDescription'),
-    url: 'https://developers.google.com/privacy-sandbox/3pcd/storage-access-api',
-    storyUrl:
-      'https://privacysandbox-stories.com/web-stories/storage-access-api/',
-    sidebarItemKey: SIDEBAR_ITEMS_KEYS.STORAGE_ACCESS,
-  },
-  {
-    title: () => I18n.getMessage('rws'),
-    description: () => I18n.getMessage('rwsDescription'),
-    url: 'https://developers.google.com/privacy-sandbox/3pcd/related-website-sets',
-    storyUrl:
-      'https://privacysandbox-stories.com/web-stories/related-website-sets/',
-    sidebarItemKey: SIDEBAR_ITEMS_KEYS.RELATED_WEBSITE_SETS,
-  },
-  {
-    title: () => 'FedCM',
-    description: () =>
-      'The Federated Credential Management API enables privacy-preserving identity federation.',
-    url: 'https://developers.google.com/privacy-sandbox/3pcd/fedcm',
-    storyUrl:
-      'https://privacysandbox-stories.com/web-stories/federated-credential-management-api/',
-    sidebarItemKey: SIDEBAR_ITEMS_KEYS.FEDERATED_CREDENTIAL,
-  },
-];
+/**
+ * Internal dependencies.
+ */
+import { SITE_BOUNDARIES_CONTENT } from '../constants';
 
 const SiteBoundaries = () => {
   return (
@@ -73,7 +31,7 @@ const SiteBoundaries = () => {
       <LandingPageContainer
         title={I18n.getMessage('siteBoundaries')}
         contentPanelTitle={I18n.getMessage('siteBoundariesDescription')}
-        content={content}
+        content={SITE_BOUNDARIES_CONTENT}
         titleStyles="text-blue-600"
         extraClasses="min-h-[78vh] w-full"
       />
